update: refatorando rota de usuario
